codes/requirements.txt
alembic==1.7.7
anyio==3.6.1
asgiref==3.5.2
asyncpg==0.25.0
backoff==2.0.1
bcrypt==3.2.2
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
databases==0.5.5
fastapi==0.78.0
fastapi-utils==0.2.1
greenlet==1.1.2
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.3
importlib-metadata==4.11.4
importlib-resources==5.7.1
Mako==1.2.0
MarkupSafe==2.1.1
passlib==1.7.4
psycopg2==2.9.3
psycopg2-binary==2.9.3
pycparser==2.21
pydantic==1.9.1
PyJWT==2.4.0
pytest==7.1.2
pytest-asyncio==0.18.3
python-decouple==3.6
requests==2.27.1
rfc3986==1.5.0
sniffio==1.2.0
SQLAlchemy==1.4.36
starlette==0.19.1
typing_extensions==4.2.0
urllib3==1.26.9
uvicorn==0.17.6
xmltodict==0.13.0
zipp==3.8.0
